+++ crates/ruma-client-api/src/threads.rs
@@ -1,9 +1,8 @@
 //! Endpoints for querying threads in a room.
 
-pub mod get_threads;
-
 #[cfg(feature = "unstable-msc4306")]
 pub mod get_thread_subscription;
+pub mod get_threads;
 #[cfg(feature = "unstable-msc4306")]
 pub mod subscribe_thread;
 #[cfg(feature = "unstable-msc4306")]